requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==0.20.0
requests==2.26.0
gspread==6.1.2
oauth2client